yts/pyproject.toml
[tool.poetry]
name = "yts"
version = "0.1.0"
description = ""
authors = ["vdmitrii"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aerich = "^0.7.1"
asyncpg = "^0.27.0"
fastapi = "^0.91.0"
gunicorn = "^20.1.0"
tortoise-orm = "^0.19.3"
uvicorn = "^0.20.0"
requests = "^2.28.2"
streamlit = "^1.22.0"
youtube-transcript-api = "^0.6.0"
openai = "^0.27.7"
python-dotenv = "^1.0.0"
tenacity = "^8.2.2"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.0"
httpx = "^0.23.3"
pre-commit = "^3.3.2"

[tool.isort]

[tool.black]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "-ra -q"
]

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
